feat: cartItem 컨트롤러 구현 및 인수 테스트 작성

+++ src/main/java/woowacourse/shoppingcart/ui/CartItemController.java
@@ -1,20 +1,23 @@
 package woowacourse.shoppingcart.ui;
 
-import java.net.URI;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+import woowacourse.auth.support.AuthenticationPrincipal;
 import woowacourse.shoppingcart.application.CartItemService;
+import woowacourse.shoppingcart.domain.Customer;
+import woowacourse.shoppingcart.dto.cart.CartItemCountRequest;
 import woowacourse.shoppingcart.dto.cart.CartItemCreateRequest;
 import woowacourse.shoppingcart.dto.cart.CartItemDto;
+import woowacourse.shoppingcart.exception.ForbiddenAccessException;
 
 @RestController
 @RequestMapping("/api/customers/{customerId}/carts")
@@ -27,26 +30,43 @@ public class CartItemController {
     }
 
     @GetMapping
-    public ResponseEntity<List<CartItemDto>> getCartItems(@PathVariable final Long customerId) {
+    public ResponseEntity<List<CartItemDto>> getCartItems(@AuthenticationPrincipal final Customer customer,
+                                                          @PathVariable final Long customerId) {
+        validateAuthorizedUser(customerId, customer);
         return ResponseEntity.ok().body(cartService.findCartsByCustomerId(customerId));
     }
 
     @PostMapping
     public ResponseEntity<Void> addCartItem(@RequestBody final CartItemCreateRequest request,
-                                            @PathVariable final Long customerId) {
-        final Long cartId = cartService.addCartItem(customerId, request);
-        final URI responseLocation = ServletUriComponentsBuilder
-                .fromCurrentRequest()
-                .path("/{cartId}")
-                .buildAndExpand(cartId)
-                .toUri();
-        return ResponseEntity.created(responseLocation).build();
+                                            @PathVariable final Long customerId,
+                                            @AuthenticationPrincipal final Customer customer) {
+        validateAuthorizedUser(customerId, customer);
+        cartService.addCartItem(customerId, request);
+        return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping
     public ResponseEntity<Void> deleteCartItem(@PathVariable final Long customerId,
-                                               @RequestParam final Long productId) {
+                                               @RequestParam final Long productId,
+                                               @AuthenticationPrincipal final Customer customer) {
+        validateAuthorizedUser(customerId, customer);
         cartService.deleteCartItem(customerId, productId);
         return ResponseEntity.noContent().build();
+    }
+
+    @PatchMapping
+    public ResponseEntity<Void> updateCartItemCount(@PathVariable final Long customerId,
+                                                    @RequestParam final Long productId,
+                                                    @RequestBody final CartItemCountRequest request,
+                                                    @AuthenticationPrincipal final Customer customer) {
+        validateAuthorizedUser(customerId, customer);
+        cartService.updateCount(customerId, productId, request.getCount());
+        return ResponseEntity.ok().build();
+    }
+
+    private void validateAuthorizedUser(long id, Customer customer) {
+        if (!customer.getId().equals(id)) {
+            throw new ForbiddenAccessException();
+        }
     }
 }
